+++ contracts/interfaces/IPoolFactory.sol
@@ -8,7 +8,7 @@ struct Params {
     bytes32 oracle; // 1bit QTI, 31bit reserve, 32bit WINDOW, ... PAIR ADDRESS
     address reserveToken;
     address recipient;
-    uint224 mark;
+    uint mark;
     uint32  initTime;
     uint    halfLife;
     uint k;
